requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asttokens==2.4.1
bcrypt==4.0.1
beautifulsoup4==4.11.1
bootstrap4==0.1.0
boto3==1.34.21
botocore==1.34.21
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
cryptography==41.0.7
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
dj-rest-auth==4.0.0
Django==5.0
django-allauth==0.54.0
django-appconf==1.0.5
django-bootstrap==0.2.4
django-bootstrap4==22.3
django-cors-headers==4.3.1
django-crispy-forms==1.14.0
django-import-export==3.3.3
django-jazzmin==2.6.0
django-redis==5.3.0
django-select2==8.1.2
django-storages==1.13.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
executing==2.0.1
firebase-admin==6.2.0
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-status==1.59.3
gunicorn==21.2.0
httplib2==0.22.0
idna==3.4
import-export==0.3.1
ipython==8.18.1
jedi==0.19.1
jmespath==1.0.1
MarkupPy==1.14
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
msgpack==1.0.7
mysql-connector-python==8.1.0
numpy==1.26.2
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
parso==0.8.3
prompt-toolkit==3.0.41
proto-plus==1.22.3
protobuf==4.21.12
psycopg2-binary==2.9.5
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
PyJWT==2.7.0
pyparsing==3.1.1
pypng==0.20220715.0
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.7
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==69.0.3
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.4
stack-data==0.6.3
tablib==3.5.0
traitlets==5.14.0
typing_extensions==4.9.0
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.7
wcwidth==0.2.12
whitenoise==6.4.0
xlrd==2.0.1
xlwt==1.3.0
